Store layout in Text object, Update comments

+++ engine/include/xng/gui/textrenderer.hpp
@@ -26,6 +26,12 @@
 #include "xng/render/2d/renderer2d.hpp"
 
 namespace xng {
+    /**
+     * Known Issues: Because the text renderer returns the rendered text as an Image there can be problems
+     * handling the text in real time as the text image size increases because there are currently several cpu operations
+     * required to move the text image between the cpu and gpu eg. swapping the columns on download or blitting
+     * the text image into an aligned texture atlas image.
+     */
     class XENGINE_EXPORT TextRenderer {
     public:
         TextRenderer() = default;
@@ -38,7 +44,7 @@ namespace xng {
 
         TextRenderer &operator=(const TextRenderer &other) = default;
 
-        Vec2f getSize(const std::string &text, const TextLayout &layout);
+        Vec2i getSize(const std::string &text, const TextLayout &layout);
 
         Text render(const std::string &text, const TextLayout &layout);
